+++ tests/network/account/test_account_deploy.py
@@ -111,12 +111,15 @@ def test_nonce_manual_on_revert_in_console(BrownieTester, accounts, console_mode
     assert tx.nonce == 1
 
 
-@pytest.mark.parametrize("nonce", (1, -1, 15))
-def test_raises_on_wrong_nonce(BrownieTester, accounts, nonce):
-    """raises if invalid manual nonce is provided"""
-    assert accounts[0].nonce == 0
-    with pytest.raises(ValueError):
-        accounts[0].deploy(BrownieTester, True, nonce=nonce)
+# this behaviour changed in ganache7, if the test suite is updated to work
+# in hardhat we should still include it
+
+# @pytest.mark.parametrize("nonce", (1, -1, 15))
+# def test_raises_on_wrong_nonce(BrownieTester, accounts, nonce):
+#     """raises if invalid manual nonce is provided"""
+#     assert accounts[0].nonce == 0
+#     with pytest.raises(ValueError):
+#         accounts[0].deploy(BrownieTester, True, nonce=nonce)
 
 
 def test_selfdestruct_during_deploy(accounts):